requirements.txt
asgiref==3.4.1
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///opt/concourse/worker/volumes/live/976f8942-f51d-4f0e-7352-2a10f0820d0e/volume/cffi_1625814703974/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///opt/concourse/worker/volumes/live/d5dda287-c0b3-4861-7262-fab05baa64dc/volume/cryptography_1616769284011/work
Django==3.2.7
djangorestframework==3.12.4
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
numpy==1.21.2
pandas==1.3.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT @ file:///opt/concourse/worker/volumes/live/bd094316-1935-4bf0-5028-b889d4e7967c/volume/pyjwt_1619682501859/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///opt/concourse/worker/volumes/live/112288ac-9cb0-4e73-768b-13baf4ca6419/volume/pysocks_1605305820043/work
python-dateutil==2.8.2
pytz==2021.1
pyupbit==0.2.19
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse==0.4.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
websockets==10.0
